Refine tmux and vim pane navigation
tmux-nav-pane now passes through the key if the pane movement is
a no-op.  This is primarily to allow C-k to continue to work in the
common case of a vertical split shell window with no upper or lower
neighbors.

tmux-wipe-pane now recognizes "ssh" panes as being shells.  This is
a bit risky, as some other app (e.g. vim) could be running there, but
then this whole C-{h,j,k,l} navigation scheme is one big house of cards.