Update vimpager to rkitover/vimpager@6f85485
This updates to the latest upstream master, in failed hopes of dealing
with two ongoing vimpager issues.  First, vimcat sometimes
(deterministically) craps out on some files when its output tempfile
doesn't exist.  Second, sometimes vimpager fails to exit with an
unadorned 'q', and requires ':q' as if it weren't a pager session.